Add files via upload
The IPFSCaller contract is a Solidity contract for storing data on IPFS (InterPlanetary File System). The contract has a single public function storeDataOnIPFS which takes a string of data as input. The function then calls the private function addDataToIPFS which sends the data to IPFS and returns the hash of the stored data. The hash is emitted through the LogIPFSHash event. The code for sending the data to IPFS is just an example and not a working implementation.